+++ src/components/Dropzone.ts
@@ -14,7 +14,7 @@ export default class DropzoneComponent {
    * Initialization.
    */
   constructor(container: HTMLElement|JQuery, options: DropzoneOption) {
-    // Check the argument.
+    // Check parameters.
     if (container instanceof $)
       container = (container as JQuery).get(0) as HTMLElement;
     else if (!(container instanceof HTMLElement))