+++ src/components/ImageInput.ts
@@ -1,6 +1,6 @@
 import hbs from 'handlebars-extd';
 import fusion from 'deep-fusion';
-import compareImg from 'compare-img';
+import compare from 'compare-img';
 import initTooltip from '~/components/initTooltip';
 import fetchDataUrl from '~/http/fetchDataUrl';
 import ImageInputOption from '~/interfaces/ImageInputOption';
@@ -18,20 +18,21 @@ import getExtensionFromDataUrl from '~/misc/getExtensionFromDataUrl';
  *   current: 'current.png',
  *   default: 'default.png'
  * });
- * imageInput.onchange(dataUrl => {});
+ * imageInput.onChange(currentImage => {});
  * 
  * @see {@link https://preview.keenthemes.com/metronic8/demo1/documentation/forms/image-input.html} Custom Bootstrap Image Input with Preview Component by Keenthemes.
  */
-export default class {
+export default class ImageInput {
   #imageInput: typeof window.KTImageInput;
-  #changeHandler: (dataUrl: string|null) => void = (dataUrl: string|null) => {};
-  #imgDataUrl: string|undefined;
+  #changeHandler: (currentImage: string|null) => void = (currentImage: string|null) => {};
+  #image: string|null = null;
+  #hiddenEl: HTMLInputElement|null = null;
 
   /**
    * Initialization.
    */
   constructor(context: HTMLDivElement|JQuery, options: ImageInputOption) {
-    // Check the argument.
+    // Check parameters.
     if (context instanceof $)
       context = (context as JQuery).get(0) as HTMLDivElement;
     else if (!(context instanceof HTMLDivElement))
@@ -48,80 +49,71 @@ export default class {
       cancelable: false,
       accept: '.png,.jpg,.jpeg,.svg',
       language: {
-        change: '変更する。',
-        remove: '削除する。',
-        cancel: '変更を取り消す。'
+        change: 'Change',
+        remove: 'Delete',
+        cancel: 'Cancel change'
       }
     }, options);
+
+    // Save hidden elements as members.
+    if (options.hiddenEl)
+      this.#hiddenEl = options.hiddenEl;
 
     // Render the current and default images.
     (async () => {
       // If the default image is not a DataURL but a URL, it is loaded remotely.
-      let defaultImg;
+      let defaultImage;
       if (options.default)
-        defaultImg = isDataUrl(options.default) ? options.default : await fetchDataUrl(options.default);
+        defaultImage = isDataUrl(options.default) ? options.default : await fetchDataUrl(options.default);
 
       // If the current image is not a DataURL but a URL, it is loaded remotely.
-      let currentImg;
       if (options.current)
-        currentImg = isDataUrl(options.current) ? options.current : await fetchDataUrl(options.current);
+        this.#image = isDataUrl(options.current) ? options.current : await fetchDataUrl(options.current);
 
       // If the current and the default image are the same, the image change cancel button is not displayed in the initial display.
-      if (options.current && options.default && await compareImg(options.current, options.default))
+      if (options.current && options.default && await compare(options.current, options.default))
         options.current = undefined;
 
-      // Set the current image to the hidden element.
-      if (options.hiddenEl && (defaultImg || currentImg))
-        // The currernt optional image takes precedence over default.
-        if (currentImg)
-          options.hiddenEl.value = currentImg as string;
-        else 
-          options.hiddenEl.value = defaultImg as string;
+      // If the current image and default image are specified, priority is given to saving and displaying the current image.
+      if (defaultImage || this.#image) {
+        // Set images for hidden elements.
+        if (this.#hiddenEl)
+          this.#hiddenEl.value = this.#image || defaultImage as string;
 
-      // Set the data URL of the current image.
-      if (defaultImg || currentImg)
-        // The currernt optional image takes precedence over default.
-        if (currentImg)
-          this.#imgDataUrl = currentImg as string;
-        else 
-          this.#imgDataUrl = defaultImg as string;
+        // If no image is currently available, the default image is saved.
+        if (!this.#image)
+          this.#image = defaultImage as string;
+      }
 
       // Rendering image input UI.
       this.#imageInput = this.#render(context as HTMLDivElement, options);
 
       // Initialize event handler.
-      this.#initEventListeners(options, defaultImg);
+      this.#initEventListeners(options, defaultImage);
 
       // Initialize Observer.
-      this.#initObserver(options.hiddenEl);
+      this.#initObserver();
     })();
   }
-
 
   /**
    * Set change event handler.
    */
-  onchange(handler: (dataUrl: string|null) => void) {
+  onChange(handler: (currentImage: string|null) => void): ImageInput {
     this.#changeHandler = handler;
-  }
-
-  /**
-   * Get the data URL of the current image.
-   */
-  getImgDataUrl(): string|undefined {
-    return this.#imgDataUrl;
+    return this;
   }
 
   /**
    * Download the current image.
    */
   download(): void {
-    if (!this.#imgDataUrl)
+    if (!this.#image)
       return;
     let link = document.createElement('a') as HTMLAnchorElement;
-    const extension = getExtensionFromDataUrl(this.#imgDataUrl);
+    const extension = getExtensionFromDataUrl(this.#image);
     link.download = `image.${extension}`;
-    link.href = this.#imgDataUrl as string;
+    link.href = this.#image as string;
     document.body.appendChild(link);
     link.click();
     document.body.removeChild(link);
@@ -136,9 +128,24 @@ export default class {
   }
 
   /**
+   * Get the data URL of the current image.
+   */
+  getImage(): string|null {
+    return this.#image;
+  }
+
+  /**
+   * Returns the hidden field of the selected image input instance.
+   */
+  getHiddenElement(): HTMLInputElement|null {
+    return this.#hiddenEl;
+  }
+
+  /**
    * Rendering image input UI.
    */
   #render(context: HTMLDivElement, options: ImageInputOption) {
+    // Context styling.
     context.classList.add(
       'image-input',
       'image-input-outline',
@@ -148,56 +155,67 @@ export default class {
     );
     // context.style.backgroundColor = '#babbbe';
     context.style.width = 'fit-content';
-    context.style.backgroundImage = `url(${options.default})`;
+
+    // If there is a default image, set it as the context background.
+    if (options.default)
+      context.style.backgroundImage = `url(${options.default})`;
     context.setAttribute('data-kt-image-input', 'false');
+
+    // Render image input HTML to context.
     context.insertAdjacentHTML('afterbegin', hbs.compile(
-      `<!--begin::Preview existing avatar-->
+      `<!--begin::Preview-->
       <div
-        {{#unless readonly}}
+        {{#unless options.readonly}}
           data-bs-toggle="tooltip"
           data-bs-dismiss="click"
-          title="{{language.change}}"
+          title="{{options.language.change}}"
         {{/unless}}
-        class="image-input-wrapper bgi-position-center {{#unless readonly}}cursor-pointer{{/unless}}"
-        style="background-size: contain; {{#if current}}background-image: url({{current}});{{/if}} width: {{width}}px; height: {{height}}px;">
+        class="image-input-wrapper bgi-position-center {{#unless options.readonly}}cursor-pointer{{/unless}}"
+        style="background-size: contain; {{#if (notEmpty options.current)}}background-image: url({{options.current}});{{/if}} width: {{options.width}}px; height: {{options.height}}px;">
       </div>
-      <!--end::Preview existing avatar-->
+      <!--end::Preview-->
       <!--begin::Edit-->
       <label
-        class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx readonly 'd-none'}}"
+        class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx options.readonly 'd-none'}}"
         data-kt-image-input-action="change"
         data-bs-toggle="tooltip"
         data-bs-dismiss="click"
-        title="{{language.change}}">
+        title="{{options.language.change}}">
         <i class="bi bi-pencil-fill fs-6"></i>
-        <input type="file" name="avatar" accept="{{accept}}" />
+        <input type="file" name="avatar" accept="{{options.accept}}" />
         <input type="hidden" name="avatar_remove" />
       </label>
       <!--end::Edit-->
-      {{#if cancelable}}
+      {{#if options.cancelable}}
         <!--begin::Cancel-->
         <span
-          class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx readonly 'd-none'}}"
+          class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx options.readonly 'd-none'}}"
           data-kt-image-input-action="cancel"
           data-bs-toggle="tooltip"
           data-bs-dismiss="click"
-          title="{{language.cancel}}">
+          title="{{options.language.cancel}}">
           <i class="bi bi-x fs-3"></i>
         </span>
         <!--end::Cancel-->
       {{/if}}
       <!--begin::Remove-->
       <span
-        class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx readonly 'd-none'}}"
+        class="btn btn-icon btn-circle btn-color-muted btn-active-color-primary w-25px h-25px bg-body shadow {{ifx options.readonly 'd-none'}}"
         data-kt-image-input-action="remove"
         data-bs-toggle="tooltip"
         data-bs-dismiss="click"
-        title="{{language.remove}}">
+        title="{{options.language.remove}}">
         <i class="bi bi-x fs-3"></i>
       </span>
-      <!--end::Remove-->`)(options));
+      <!--end::Remove-->`)({options}));
+
+    // Initialize tooltip.
     initTooltip(context);
+
+    // Initialize the image input component.
     const imageInput = new window.KTImageInput(context);
+
+    // If there is currently an image and the Cancel button is disabled, the Delete Image button is shown.
     if (options.current && !options.cancelable)
       imageInput.removeElement.style.display = 'flex';
 
@@ -211,28 +229,40 @@ export default class {
   /**
    * Initialize event handler.
    */
-  #initEventListeners(options: ImageInputOption, defaultImg: string|undefined): void {
+  #initEventListeners(options: ImageInputOption, defaultImage: string|undefined): void {
+    // This event fires on when the image input has been changed.
     this.#imageInput.on('kt.imageinput.changed', async (input: typeof window.KTImageInput) => {
       // If cancellation is disabled, the delete button is forcibly displayed when editing the image.
       if (!options.cancelable)
         input.removeElement.style.display = 'flex';
     });
+
+    // This event fires on when the image input has been removed.
     this.#imageInput.on('kt.imageinput.removed', async (input: typeof window.KTImageInput) => {
-      if (defaultImg) {
-        // Set the default image when the image is canceled.
-        if (options.hiddenEl)
-          options.hiddenEl.value = defaultImg as string;
+      if (defaultImage) {
+        // If there is a default image.
+        // Set default image for hidden elements.
+        if (this.#hiddenEl)
+          this.#hiddenEl.value = defaultImage as string;
 
         // Set the data URL of the current image.
-        this.#imgDataUrl = defaultImg as string;
+        this.#image = defaultImage as string;
+      } else {
+        // If there is no default image.
+        // Clear hidden elements.
+        if (this.#hiddenEl)
+          this.#hiddenEl.value = '';
 
-        // Invoke change event.
-        this.#changeHandler(defaultImg||null);
+        // Currently clearing the image.
+        this.#image = null;
       }
 
       // If cancellation is disabled, the delete button is forcibly hidden when deleting an image.
       if (!options.cancelable)
         input.removeElement.style.display = 'none';
+
+      // Invoke change event.
+      this.#changeHandler(this.#image);
     });
 
     // Unless it is read-only.
@@ -250,37 +280,29 @@ export default class {
   /**
    * Initialize Observer.
    */
-  #initObserver(hiddenEl: HTMLInputElement|undefined) {
+  #initObserver() {
     const observer = new MutationObserver(async mutationsList => {
       for (let mutation of mutationsList) {
+        // If it is not a style or attribute that has been changed, do nothing.
         if (mutation.type !== 'attributes' || mutation.attributeName !== 'style')
           continue;
 
-        // Get the image source from the background style.
+        // Get the source of the image.
         const matches = this.#imageInput.wrapperElement.style.backgroundImage.match(/(?:\(['"]?)(.*?)(?:['"]?\))/);
         if (!matches)
           continue;
-        const backgroundImg = matches[1];
 
-        // Active image.
-        const activeImg = isDataUrl(backgroundImg) ? backgroundImg : await fetchDataUrl(backgroundImg);
+        // Get current image.
+        this.#image = isDataUrl(matches[1]) ? matches[1] : await fetchDataUrl(matches[1]);
 
-        // Set the selected image to the Hidden element.
-        if (hiddenEl)
-          hiddenEl.value = activeImg;
-
-        // Set the data URL of the current image.
-        this.#imgDataUrl = activeImg;
+        // Update on hidden elements.
+        if (this.#hiddenEl)
+          this.#hiddenEl.value = this.#image as string;
 
         // Invoke change event.
-        this.#changeHandler(activeImg);
+        this.#changeHandler(this.#image);
       }
     });
-    observer.observe(this.#imageInput.wrapperElement, { 
-      attributes: true, 
-      childList: false,
-      characterData: true,
-      attributeFilter: ['style'],
-    });
+    observer.observe(this.#imageInput.wrapperElement, {attributes: true, childList: false, characterData: true, attributeFilter: ['style']});
   }
 }