+++ src/components/initTooltip.ts
@@ -4,7 +4,7 @@ import bootstrap from '~/misc/bootstrap';
  * Initialize the tooltip.
  */
 export default (context: JQuery|HTMLElement, tooltipSelector: string = '[data-bs-toggle="tooltip"]'): bootstrap.Tooltip[] => {
-  // Check the argument.
+  // Check parameters.
   if (context instanceof HTMLElement)
     context = $(context);
   else if (!(context instanceof $))